refactor: ProductResponse 생성
